pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qrcode-api"
version = "0.1.0"
description = "API de génération de QR codes stateless"
authors = [
    { name = "Spec Windsurf", email = "spec@windsurf.dev" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "qrcode==7.4.2",
    "Pillow==10.1.0",
    "PyPDF2==3.0.1",
    "python-multipart==0.0.9",
    "python-jose==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "slowapi==0.1.8",
    "python-dotenv==1.0.0",
    "httpx==0.26.0"
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = "src/"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = ["src/__pycache__", "tests/__pycache__", ".venv", "env"]

[tool.mypy]
python_version = 3.11
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
files = ["src", "tests"]
